requirements.txt
streamlit
pillow
pandas
matplotlib
openpyxl
